+++ src/static/js/mapMultipleTracks.js
@@ -3,6 +3,8 @@ document.addEventListener('DOMContentLoaded', function()
     initMap();
 });
 
+const PATTERN_TRACK_NAME = /(\d{4}-\d{2}-\d{2} - .*)<\/a>/;
+
 function initMap()
 {
     let map = initMapBase();
@@ -41,9 +43,39 @@ function initMap()
                     const trackId = fallbackName;
                     const gpxInfoForTrack = getGpxInfoById(trackId);
 
-                    geojson.name = '<a href="' + gpxInfoForTrack.trackUrl+ '">' + gpxInfoForTrack.trackName + '</a>'
-                    this._loadRoute(geojson);
+                    geojson.name = '<a href="' + gpxInfoForTrack.trackUrl + '">' + gpxInfoForTrack.trackName + '</a>'
+                    this._loadRoute(geojson, gpxInfoForTrack.gpxUrl);
                 }
+            },
+            _loadRoute: function(data, gpxUrl)
+            {
+                if(!data)
+                {
+                    return;
+                }
+
+                var line_style = {
+                    color: this._uniqueColors(this._tracks.length)[this._tracks.indexOf(gpxUrl)],  // access color by real index in tracks array instead of count of loaded elements
+                    opacity: 0.75,
+                    weight: 5,
+                    distanceMarkers: this.options.distanceMarkers_options,
+                };
+
+                var route = L.geoJson(data, {
+                    name: data.name || '',
+                    style: (feature) => line_style,
+                    distanceMarkers: line_style.distanceMarkers,
+                    originalStyle: line_style,
+                    filter: feature => feature.geometry.type != "Point",
+                });
+
+                this._elevation.import(this._elevation.__LGEOMUTIL).then(() =>
+                {
+                    route.addTo(this._layers);
+
+                    route.eachLayer((layer) => this._onEachRouteLayer(route, layer));
+                    this._onEachRouteLoaded(route);
+                });
             },
         });
 
@@ -63,6 +95,22 @@ function initMap()
             legend: true,
             distanceMarkers: false
         });
+
+        routes._legend.options.sortLayers = true
+        routes._legend.options.sortFunction = function(layerA, layerB, nameA, nameB)
+        {
+            let realNameA = PATTERN_TRACK_NAME.exec(nameA)[1];
+            let realNameB = PATTERN_TRACK_NAME.exec(nameB)[1];
+            if(realNameA < realNameB)
+            {
+                return 1;
+            }
+            if(realNameA > realNameB)
+            {
+                return -1;
+            }
+            return 0;
+        };
 
         routes.addTo(map);